+++ crates/proto/src/domain/merkle.rs
@@ -1,6 +1,8 @@
 use miden_objects::{
-    Digest, Word,
-    crypto::merkle::{LeafIndex, MerklePath, MmrDelta, SmtLeaf, SmtProof},
+    Word,
+    crypto::merkle::{
+        Forest, LeafIndex, MerklePath, MmrDelta, SmtLeaf, SmtProof, SparseMerklePath,
+    },
 };
 
 use super::{convert, try_convert};
@@ -29,7 +31,43 @@ impl TryFrom<&proto::merkle::MerklePath> for MerklePath {
     type Error = ConversionError;
 
     fn try_from(merkle_path: &proto::merkle::MerklePath) -> Result<Self, Self::Error> {
-        merkle_path.siblings.iter().map(Digest::try_from).collect()
+        merkle_path.siblings.iter().map(Word::try_from).collect()
+    }
+}
+
+impl TryFrom<proto::merkle::MerklePath> for MerklePath {
+    type Error = ConversionError;
+
+    fn try_from(merkle_path: proto::merkle::MerklePath) -> Result<Self, Self::Error> {
+        (&merkle_path).try_into()
+    }
+}
+
+// SPARSE MERKLE PATH
+// ================================================================================================
+
+impl From<SparseMerklePath> for proto::merkle::SparseMerklePath {
+    fn from(value: SparseMerklePath) -> Self {
+        let (empty_nodes_mask, siblings) = value.into_parts();
+        proto::merkle::SparseMerklePath {
+            empty_nodes_mask,
+            siblings: siblings.into_iter().map(proto::digest::Digest::from).collect(),
+        }
+    }
+}
+
+impl TryFrom<proto::merkle::SparseMerklePath> for SparseMerklePath {
+    type Error = ConversionError;
+
+    fn try_from(merkle_path: proto::merkle::SparseMerklePath) -> Result<Self, Self::Error> {
+        Ok(SparseMerklePath::from_parts(
+            merkle_path.empty_nodes_mask,
+            merkle_path
+                .siblings
+                .into_iter()
+                .map(Word::try_from)
+                .collect::<Result<Vec<_>, _>>()?,
+        )?)
     }
 }
 
@@ -39,7 +77,10 @@ impl TryFrom<&proto::merkle::MerklePath> for MerklePath {
 impl From<MmrDelta> for proto::mmr::MmrDelta {
     fn from(value: MmrDelta) -> Self {
         let data = value.data.into_iter().map(proto::digest::Digest::from).collect();
-        proto::mmr::MmrDelta { forest: value.forest as u64, data }
+        proto::mmr::MmrDelta {
+            forest: value.forest.num_leaves() as u64,
+            data,
+        }
     }
 }
 
@@ -48,10 +89,10 @@ impl TryFrom<proto::mmr::MmrDelta> for MmrDelta {
 
     fn try_from(value: proto::mmr::MmrDelta) -> Result<Self, Self::Error> {
         let data: Result<Vec<_>, ConversionError> =
-            value.data.into_iter().map(Digest::try_from).collect();
+            value.data.into_iter().map(Word::try_from).collect();
 
         Ok(MmrDelta {
-            forest: value.forest as usize,
+            forest: Forest::new(value.forest as usize),
             data: data?,
         })
     }
@@ -74,12 +115,13 @@ impl TryFrom<proto::smt::SmtLeaf> for SmtLeaf {
                 Ok(Self::new_empty(LeafIndex::new_max_depth(leaf_index)))
             },
             proto::smt::smt_leaf::Leaf::Single(entry) => {
-                let (key, value): (Digest, Word) = entry.try_into()?;
+                let (key, value): (Word, Word) = entry.try_into()?;
 
                 Ok(SmtLeaf::new_single(key, value))
             },
             proto::smt::smt_leaf::Leaf::Multiple(entries) => {
-                let domain_entries: Vec<(Digest, Word)> = try_convert(entries.entries)?;
+                let domain_entries: Vec<(Word, Word)> =
+                    try_convert(entries.entries).collect::<Result<_, _>>()?;
 
                 Ok(SmtLeaf::new_multiple(domain_entries)?)
             },
@@ -95,7 +137,7 @@ impl From<SmtLeaf> for proto::smt::SmtLeaf {
             SmtLeaf::Empty(leaf_index) => Leaf::Empty(leaf_index.value()),
             SmtLeaf::Single(entry) => Leaf::Single(entry.into()),
             SmtLeaf::Multiple(entries) => {
-                Leaf::Multiple(proto::smt::SmtLeafEntries { entries: convert(entries) })
+                Leaf::Multiple(proto::smt::SmtLeafEntries { entries: convert(entries).collect() })
             },
         };
 
@@ -106,11 +148,11 @@ impl From<SmtLeaf> for proto::smt::SmtLeaf {
 // SMT LEAF ENTRY
 // ------------------------------------------------------------------------------------------------
 
-impl TryFrom<proto::smt::SmtLeafEntry> for (Digest, Word) {
+impl TryFrom<proto::smt::SmtLeafEntry> for (Word, Word) {
     type Error = ConversionError;
 
     fn try_from(entry: proto::smt::SmtLeafEntry) -> Result<Self, Self::Error> {
-        let key: Digest = entry
+        let key: Word = entry
             .key
             .ok_or(proto::smt::SmtLeafEntry::missing_field(stringify!(key)))?
             .try_into()?;
@@ -123,8 +165,8 @@ impl TryFrom<proto::smt::SmtLeafEntry> for (Digest, Word) {
     }
 }
 
-impl From<(Digest, Word)> for proto::smt::SmtLeafEntry {
-    fn from((key, value): (Digest, Word)) -> Self {
+impl From<(Word, Word)> for proto::smt::SmtLeafEntry {
+    fn from((key, value): (Word, Word)) -> Self {
         Self {
             key: Some(key.into()),
             value: Some(value.into()),